temperature format changed

+++ app/src/main/java/com/example/test_task_weather_forecast/data/model/WeatherModel.kt
@@ -11,7 +11,7 @@ data class WeatherModel(
 
 data class WeatherForecastHourly(
     val dtTxt: String,
-    val temp: Double,
+    val temp: Int,
     val weather: List<WeatherHourly>,
 )
